requirements.txt
streamlit
langgraph
langchain
sqlite-utils
pandas
google-generativeai
pytesseract
pillow
opencv-python
plotly
pydeck
geopy
python-dotenv
requests
scikit-learn
bcrypt
fpdf
langfuse